requirements.txt
altgraph==0.15
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
cycler==0.10.0
DateTime==4.2
future==0.16.0
idna==2.6
itchat==1.3.10
jieba==0.39
macholib==1.9
matplotlib==2.1.2
numpy==1.14.0
pefile==2017.11.5
Pillow==5.0.0
PyInstaller==3.3.1
pyparsing==2.2.0
pypiwin32==220
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.6.1
pyttsx3==2.7
pytz==2017.3
requests==2.18.4
six==1.11.0
tomd==0.1.3
urllib3==1.22
wordcloud==1.3.3
zope.interface==4.4.3
